Add Core Heat mechanism to Planetary ISRU
Hi,

This changes the Planetary ISRU to use the same core heat/thermal efficiency system that the stock ISRU uses. I've tested the change, and it works just like I expected for a simple setup. I actually just pasted in the code from the stock ISRU's config file, so the heat generation specs are the same:

- 200 kW generated per converter module at 1000K (shown as "required cooling" in VAB)
- 500 kW of max usable cooling
- shuts down at 4000K, but the heat-generation curve falls off so sharply as temperature rises that it's actually not easy to get it that hot (radiative cooling eventually catches it, in simple cases)

My only caveat is that as I have it here, this part generates 100% of the heat but works at 80% the rate of the stock ISRU. Would it make sense to lower the heat generation a bit, for balance purposes? FWIW, this part + one planetary drill should require a total of 250kW cooling at peak efficiency, which is a convenient figure.

Anyway, let me know if this is helpful. I figured that I should actually learn to use GitHub and do this dang thing myself, instead of just pestering you on the KSP forums about it!

---Sargon