Fix CLS passability issues
* Planetary Central Hub bottom node is no longer passable - it has no hatch, and other parts have equivalent nodes already marked the same
* Garage Adapter is now passable - it clearly has hatches on both sides
* Airlock (end version) now has leg1 and leg2 nodes marked as not passable.
* Convert tabs to spaces to make the file consistent

Please note, however, that it looks like current version of CLS (1.2.3.0) seems not to process nodes defined in the `NODE{}` module. All of these nodes just inherit the passable trait. This applies to all parts that have some of their nodes defined in such way - airlock, adapter, engine, etc. The fix would be to either change the config to list all passable nodes explicitly via `passablenodes` instead of using `passable`, or actually fix it in CLS.